Medical_Parser_Project/requirements.txt
pytesseract
Pillow
google-generativeai
pydantic
python-dotenv
